+++ PDFWriter/CharStringType1Interpreter.cpp
@@ -227,7 +227,7 @@ EStatusCode CharStringType1Interpreter::InterpretNumber(Byte inBuffer,InputCharS
 			return PDFHummus::eFailure;
 
 
-		operand = (long)(
+		operand = (int32_t)(
 						((unsigned long)(byte1) << 24) + 
 						((unsigned long)(byte2) << 16) + 
 						((unsigned long)(byte3) << 8) +
